+++ tools/src/main/kotlin/com/dreampany/tools/data/source/remote/RemoteWordDataSource.kt
@@ -30,15 +30,15 @@ class RemoteWordDataSource(
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 
-    override fun getTracks(startAt: Int, limit: Int): List<String>? {
+    override fun getTracks(startAt: Long, limit: Long): List<String>? {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 
-    override fun getTracksRx(startAt: Int, limit: Int): Maybe<List<String>> {
+    override fun getTracksRx(startAt: Long, limit: Long): Maybe<List<String>> {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 
-    override fun getRawItemsByLength(id: String, limit: Int): List<String>? {
+    override fun getRawItemsByLength(id: String, limit: Long): List<String>? {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 
@@ -62,7 +62,7 @@ class RemoteWordDataSource(
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 
-    override fun getItems(limit: Int): List<Word>? {
+    override fun getItems(limit: Long): List<Word>? {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 
@@ -78,11 +78,11 @@ class RemoteWordDataSource(
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 
-    override fun getItemsRx(limit: Int): Maybe<List<Word>> {
+    override fun getItemsRx(limit: Long): Maybe<List<Word>> {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 
-    override fun getSearchItems(query: String, limit: Int): List<Word>? {
+    override fun getSearchItems(query: String, limit: Long): List<Word>? {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
